moved the findIndoorLocation from the navigationCoordinator to the mappedin service
